+++ index.js
@@ -48,10 +48,10 @@ module.exports.prototype._onAll = function (patterns, handler, context, once) {
   var self = this
   var chemicalsFound = []
   var createSingleHandler = function(index){
-    return function(c){
+    return function(c, callback){
       chemicalsFound[index] = c
       if(self.utils.isFilledArray(chemicalsFound) && chemicalsFound.length == patterns.length) {
-        handler.apply(context, chemicalsFound)
+        handler.apply(context, chemicalsFound.concat([callback]))
       }
     }
   }
@@ -202,8 +202,13 @@ module.exports.prototype._emit = function (chemical, options, callback) {
         i -= 1;
         listenersCount -= 1;
       }
-
       var aggregated = listener.handler.call(listener.context, chemical, callback || function noop () {})
+      if (listener.handler.length <= 1 && callback) {
+        // handler is accepting only a chemical
+        // but emit has been called with a callback, 
+        // so let's callback inflavor of the handler
+        callback()
+      }
       if (aggregated === true) return true // halt chemical transfer, it has been aggregated
     }
   }